# perf/remove unnecessary clone