Removes an unnecessary clone when resolving feature toggles. This offers us a nice perf improvement when using 30K toggles. 

**Baseline benchmark**

Summary:
  Total:	13.8719 secs
  Slowest:	12.5792 secs
  Fastest:	2.0131 secs
  Average:	5.9556 secs
  Requests/sec:	7.2088
  

Response time histogram:
  2.013 [1]	|■
  3.070 [3]	|■■■
  4.126 [8]	|■■■■■■■
  5.183 [17]	|■■■■■■■■■■■■■■■■
  6.240 [43]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  7.296 [23]	|■■■■■■■■■■■■■■■■■■■■■
  8.353 [1]	|■
  9.409 [0]	|
  10.466 [1]	|■
  11.523 [1]	|■
  12.579 [2]	|■■


Latency distribution:
  10% in 4.0238 secs
  25% in 5.0284 secs
  50% in 5.9740 secs
  75% in 6.9629 secs
  90% in 7.0979 secs
  95% in 8.1584 secs
  99% in 12.5792 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0011 secs, 2.0131 secs, 12.5792 secs
  DNS-lookup:	0.0003 secs, 0.0000 secs, 0.0019 secs
  req write:	0.0001 secs, 0.0000 secs, 0.0014 secs
  resp wait:	3.1577 secs, 0.5576 secs, 6.9597 secs
  resp read:	2.5922 secs, 0.0871 secs, 6.1102 secs

Status code distribution:
  [200]	100 responses



**Benchmark after changes**

Summary:
  Total:	13.2835 secs
  Slowest:	9.6335 secs
  Fastest:	1.5030 secs
  Average:	4.3143 secs
  Requests/sec:	10.6147
  

Response time histogram:
  1.503 [1]	|■
  2.316 [3]	|■■
  3.129 [5]	|■■■■
  3.942 [50]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  4.755 [51]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  5.568 [22]	|■■■■■■■■■■■■■■■■■
  6.381 [1]	|■
  7.194 [4]	|■■■
  8.007 [1]	|■
  8.820 [2]	|■■
  9.633 [1]	|■


Latency distribution:
  10% in 3.4027 secs
  25% in 3.5432 secs
  50% in 4.3285 secs
  75% in 4.7461 secs
  90% in 5.3358 secs
  95% in 6.6417 secs
  99% in 9.6335 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0005 secs, 1.5030 secs, 9.6335 secs
  DNS-lookup:	0.0003 secs, 0.0000 secs, 0.0018 secs
  req write:	0.0001 secs, 0.0000 secs, 0.0030 secs
  resp wait:	2.5236 secs, 0.6410 secs, 5.2362 secs
  resp read:	1.7900 secs, 0.0868 secs, 4.3949 secs

Status code distribution:
  [200]	141 responses